Refactors out screens into separate files
